[BI-1555] - Refactoring TraitDAO and ProgramDAO to follow the interface/impl design pattern
Introduced flag to suppress refreshing cache on startup if desired (useful for integration tests)
