+++ src/main/resources/application.yml
@@ -34,6 +34,7 @@ micronaut:
   bi:
     api:
       version: v1
+      run-scheduled-tasks: true
   security:
     enabled: true
     oauth2:
